refactor: make cli.scala a thin orchestrator
Extract formatting functions into format.scala and all command
implementations + dispatch into commands.scala, reducing cli.scala
from 1,243 lines to ~190 lines of pure arg parsing and entry point.

Also removes `private` from extractImports in extraction.scala
since cmdOverview (now in commands.scala) calls it cross-file.
